ai-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
PyYAML==6.0.1
requests==2.31.0
ultralytics